refactor: moving the datalake client
